Document modules and remove floating buzzer math

+++ include/tone_table.h
@@ -2,7 +2,13 @@
 #define TONE_TABLE_H
 
 #include <stddef.h>
+#include <stdint.h>
 
-const double *tone_table_base_frequencies(size_t *count);
+/*
+ * Provides access to the base frequencies (in Hertz) used by the buzzer module.
+ * Values are stored as integers to comply with the embedded target restrictions
+ * on floating-point usage.
+ */
+const uint16_t *tone_table_base_frequencies(size_t *count);
 
 #endif
